requirements.txt
opencv-python
mediapipe
Pillow